Update patch set 1
Patch Set 1:

(1 comment)

I tested this patch set and it didn't work for me, behaving the same as when I implemented the same feature before (no patch submitted from me on this one though since I never actually got it to work.) I'm trying to get WebPositive's back and forward shortcuts to be ctrl+left and ctrl+right instead of cmd+left and cmd+right but passing in (B_NO_MODIFIERS | B_CONTROL_KEY) to AddShortcut but this doesn't show the modifiers correctly in the menu and the shortcuts don't work either.

Do you mind if I take over this patch set X512? I've removed all the F-key support from the patch so that it can go back to using a char return value instead of using an int return value. Not that F-key modifier support is unwanted but it changes the API. I also changed B_NO_MODIFIERS to B_NO_COMMAND_KEY.

Patch-set: 1