Update patch set 1
Patch Set 1:

BTW, Personally I'd think it's generally speaking a bad idea to turn on DPMS when no mode was ever set yet using an accelerant. Setting a display mode is actually a piece of the card configuration, along with sending out timing signals to a connected monitor. When turning DPMS on, while configuration is not yet complete, it's very possible the driver/accelerant or even card hardware will fall to a hang/stall situation from which it's sometime even possible to be unable to return from. DPMS saves power (when not fully ON) in connected monitors, but (on some drivers, especially laptops I'd assume) it also might save power inside the graphics card itself (turning off PLL timings for instance).

Apparantly at this time we only see this actually happening on Intel driver, and specifically on generation 3 graphics cards.

On a sidenote: 
Of course also accelerants -could- be updated to block setting DPMS to ON when no mode was set yet. 
Note please that modes set during boot do not count here, as what happens there is unknown to the accelerant. Even it one were to relay the set mode to the accelerant from there, the internal hardware settings in use are still unknown.

Patch-set: 1