Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> BTW, Personally I'd think it's generally speaking a bad idea to turn on DPMS when no mode was ever set yet using an accelerant. 

Yes, the only reason this had been moved is that it was trying to set DPMS on at every workspace switch. And apparently with the VESA driver, some VESA BIOSes will interpret that as turning the display off, and back on again.

It seems that sending requests to turn DPMS ON when it's in fact already ON is another bad idea. So we need something that avoids both bad ideas 😊

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>