Update patch set 1
Patch Set 1:

> This works well if you make sure to set the exact same mode on all desktop (using screen preferences).  But with the Intel driver (at least, don't know about others) I had cases where workspaces would use apparently identical mode (same resolution and refresh rate) but in fact slightly different (small changes in timing, maybe?). I am not sure how I ended up in that situation, maybe after an SDL or Allegro game had changed and then restored the display resolution? And in that case I'd get a useless mode switch when switching workspaces. Re-setting the same resolution for all workspaces in screen preferences solves this, but it's a bit annoying to have to do it manually.

On a sidenote: the keyboard shortcut for resolution fallback does not sync nicely with the ScreenPrefs. Also, reverting ScreenPrefs seems to only work for the active workspace. I always have to actively 'set' it again after a revert to prevent (some of) these differences to remain between this, and the other workspaces.

Patch-set: 1