Update patch set 3
Patch Set 3:

gHtifRegs = 0 change is needed to avoid using HTIF when it is not available. gHtifRegs != NULL check is used to detect that HTIF is present. 0x40008000 is TinyEMU HTIF address. HTIF is emulator specific device that provide serial IO and shutdown capability (and maybe something else depending on virtual machine).

Patch-set: 3