Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Sphinx supports markdown, right? Is there really a reason we have to use rST? If people want to write documentation in RST that's fine, but do we have to convert everything all at once here?

rST is the default and the recommended option. Markdown is possible through a plugin but its integration is less well documented and probably has limitations. And using a single format seems like a good idea.

I already did the conversion anyway (with some manual fixups), please dont' make me revert all this hard work 😊

> It looks like the document hierarchy in the preview link you have here is a bit flat 

I have not yet created all the table-of-content pages that will provide structure. I just wanted to convert the existing docs first and get the toolchain in place. Also, some discussion will be needed as the existing structure may not work so well. For example, it probably makes sense to document the interface kit and app_server in a single chapter, as they are closely coupled together. But at least we have a starting point.

There are some obvious things to do (put kernel and drivers in a separate chapter, for example) and we can also tweak how many levels we want to show in each table of contents to help people get an overview of the document structure, when there is one.

Patch-set: 1