Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Personally I also think the 1px ones look better for the Haiku control look; the 2px ones do seem more "90s."

It's quite difficult to see the dots on the current 1px version, especially once you are at 1080p resolution or above. The 2px version fit in nicely with the HaikuControlLook aesthetic IMHO. What's wrong with the '90's? Do you have a suggestion for a better looking dots knob design?

Patch-set: 2