Update patch set 2
Patch Set 2:

Haiku's style is reminiscent of the 90s, but I like to think of it as more "timeless." We have UI elements in our design from just about every era of computing, and I think it blends well.

Maybe a hybrid approach, where one of the sides is 2px and the other is 1px, might work. I think also using a slightly more muted white, specifically the one from the top border of the button bar, would help too.

Patch-set: 2