Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Haiku's style is reminiscent of the 90s, but I like to think of it as more "timeless." We have UI elements in our design from just about every era of computing, and I think it blends well.
> 
> Maybe a hybrid approach, where one of the sides is 2px and the other is 1px, might work. I think also using a slightly more muted white, specifically the one from the top border of the button bar, would help too.

I could draw a 2x2 dark square and then draw a 1x1 light dot on top of the top-left corner of the dark square visually indicating a light source coming from the top-left corner of the screen casting a shadow around the dot. That might look a little better than the current extension of the "light" square a pixel beyond the edge of the dark square on the top left.

The light square is pure white, anything less than pure white wouldn't provide enough contrast. If we drew the light dot in the top left corner of the dark square it would make the light dot appear darker due to optical illusion of there being dark pixels next to it.

Patch-set: 2