Update patch set 2
Patch Set 2:

I'm not sure this is really the right approach. But, I have a lot of thoughts about HiDPI that I realized the current system will not handle, and some of the fixes I have in mind might make this not necessary. So, let's see what happens in discussion once I raise those

Patch-set: 2
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>