Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Not sure about this one, honestly. MIME types should make file extensions obsolete, but I guess there are plenty of circumstances where that is not the case. Anyone else have thoughts?

Extensions are not only obsolete, they are also redundant with the file icon.

There will always be problematic cases whatever we do here. Some files have important text at the start, other at the end, other in the middle. And the computer can't really guess, or even if it can (you could detect which part of the filenames has most variation?), it would be confusing to switch between different modes depending on the visible filenames.

So, I don't really mind. We have to truncate somewhere, and I'm not sure it matters where.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>