Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Not sure about this one, honestly. MIME types should make file extensions obsolete, but I guess there are plenty of circumstances where that is not the case. Anyone else have thoughts?
> 
> Extensions are not only obsolete, they are also redundant with the file icon.

I was messing with images a lot working on the Tracker thumbnail branch. Extensions for image files are not obsolete because you often have to interchange image files with other systems and the web where the extension is used to identify the type. If you have a bunch of images in a directory of different types: jpeg, png, gif, webp, ai the file icon will be the same for all (or now possibly a thumbnail.) so the extension helps here.

> 
> There will always be problematic cases whatever we do here. Some files have important text at the start, other at the end, other in the middle. And the computer can't really guess, or even if it can (you could detect which part of the filenames has most variation?), it would be confusing to switch between different modes depending on the visible filenames.
> 
> So, I don't really mind. We have to truncate somewhere, and I'm not sure it matters where.

If we truncate in the middle instead of the end it shows the extension, which is still relevant in some cases. We can only do so much to show the relevant part of the truncated name.

Patch-set: 2