Update patch set 2
Patch Set 2:

The biggest problem is that dependencies do not go through packages for whatever reason; that is, if I change one file, it rebuilds all packages and not just whatever one needs to be updated. Fixing that would probably resolve most concerns, as then the userguides et al. would only be built a single time.

Patch-set: 2