Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The biggest problem is that dependencies do not go through packages for whatever reason; that is, if I change one file, it rebuilds all packages and not just whatever one needs to be updated. Fixing that would probably resolve most concerns, as then the userguides et al. would only be built a single time.

"whatever reason" is the haiku image depends on a local package repository, which in turn depends on all packages, even those that are not included in the image.

But I'm not sure why the userguide packages get rebuilt, if none of the files changed, this should not happen.

In the case of the userguide, I would consider moving it out of the git repository and versionning it separately, not only it would fix this, but also users wouldn't have to needlessly re-download it everytime we change something in haiku. Same thing for the makefile engine.

Patch-set: 2