Update patch set 2
Patch Set 2:

> Thanks for the patch! I have found some coding style issues (unfortunately, I can't really comment on the changes otherwise).

To provide some context since the commit message doesn't give a lot of details (let me know if I get something wrong here):

Our HID parser expected that the keyboard modifiers would be reported each separatel, but other keys would be reported using a shared buffer (there is space for 8 "currently pressed" keys in the buffer usually. "Gamer" keyboards, however, support "n-key-rollover", which basically means each key gets a separate bit in the HId report and the keyboard can report any combination of pressed keys.

The way it is handled here is a bit wasteful, since we end up with a separate HIDReportItem for each bit in the report. It is modelled after what we did for the modifier keys previously.

It would be more efficient to extend the HIDReportItem class to allow for multiple range of keys reported as individual bits to be stored in a single HIDReportItem instance, but I don't know how easy it would be to do this.

Related bug report (not a lot of details there): #17189

It could possibly help with #16852, #16765, #16647, maybe #14919.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>