Update patch set 2
Patch Set 2:

> The way it is handled here is a bit wasteful, since we end up with a separate HIDReportItem for each bit in the report. It is modelled after what we did for the modifier keys previously.
> 
> It would be more efficient to extend the HIDReportItem class to allow for multiple range of keys reported as individual bits to be stored in a single HIDReportItem instance, but I don't know how easy it would be to do this.
> 

As I see it, HIDReportItem should be 1:1 to inputs, no matter the number of them. One option may be to lose some weight on this class. I think there are some properties that can be easily gathered from HIDReportItem and others can be squashed into a single integer, like boolean ones.

Antoher option perhaps, is to get rid of HIDReportItem and just query input to HIDReport, but this may require a bigger refactor on whole hid stack.

I am going to take a look to your approach, in fact was my first idea, but suddenly found this way was easier to implement.

> Related bug report (not a lot of details there): #17189
> 
> It could possibly help with #16852, #16765, #16647, maybe #14919.

There was a missing memset after malloc'ing some place for input keys on Keyboard handler, I am afraid that some of those issues are related to this.

Patch-set: 2