Update patch set 2
Patch Set 2:

> Ok, I've come with a new idea. Current code assigns one HIDReportItem to each single usage and one for the first min-max pair. My patch assigns a HIDReportItem to all usages between a min-max pair. Simple but suboptimal.
> 
> It is not a good idea to me a HIDReportItem storing multiple ranges of min-max. Seems like a perversion of Item concept. 

Yes, one item per min-max pair seems good. Otherwise it's just moving the complexity inside HIDReportItem.

Is that what happens in the current patch? It's a bit hard to see from just the diff, I need to re-read all the code to remember how this all works...

> 
> Perhaps, wouldn't be too difficult to just keep the original idea but instead of ignoring second min-max range, assign to a new HIDReportItem. This way api remains untouched and memory footprint keeps almost equal.

The memory footprint is mainly defined by the MAX_KEYS constant. In the current patch it gets a major increase (from 8 to 128), does it need to be that big? Also, there is now no difference between "keys" and "modifiers", right? So, can we store them all in a single array instead of two?

Regarding that constant, you added or modified some tests for it and it looks like if the array is full, extra items are silently ignored. I think emitting somethin to the logs (using dprintf or TRACE_ALWAYS) would be nice in that case, especially if we make it smaller than 128 and fear that it might be too small for some keyboards.

Patch-set: 2