Update patch set 2
Patch Set 2:

> Yes, one item per min-max pair seems good. Otherwise it's just moving the complexity inside HIDReportItem.
> 
> Is that what happens in the current patch? It's a bit hard to see from just the diff, I need to re-read all the code to remember how this all works...
> 

No, in fact all min-max at HIDReportItem is ignored. Min-max ranges are converted to a bunch of single usages at parser. That is the reason why I had to increase MAX_KEYS, to make enough room for all keyboard keys.

> The memory footprint is mainly defined by the MAX_KEYS constant. In the current patch it gets a major increase (from 8 to 128), does it need to be that big? Also, there is now no difference between "keys" and "modifiers", right? So, can we store them all in a single array instead of two?

Imho, HIDReportItems store too much info that is available at HIDReport, and some bools that can be squashed into a shared int8. 

There is another problem with HIDReportItems with usage ranges in bitmap mode. Data() returns 1/0, so you have to compute, stored elsewhere, the usage activated. This is however, easy to fix.
 
> Regarding that constant, you added or modified some tests for it and it looks like if the array is full, extra items are silently ignored. I think emitting somethin to the logs (using dprintf or TRACE_ALWAYS) would be nice in that case, especially if we make it smaller than 128 and fear that it might be too small for some keyboards.

Yeah, I pretended to just focus on fixing issues. I am willing to fix all of this details in a future patch series. 

Give me a couple of days to submit a patch with this new min-max approach.

Patch-set: 2