USB: Make the SuperSpeed Endpoint Companion Descriptor publicly accessible.
This also simplifies its access inside the USB stack itself.

As usb_endpoint_infos are stored in a publicly accessible array,
this is unavoidably a kernel USB ABI break. (Userspace APIs
do not use these methods or structs.)

Unfortunately, this also means we cannot keep compatibility with
the "USB2" kernel ABI, either, which relied on having identical
structs to the USB3 one. I don't know if anyone is using BeOS R5
USB kernel drivers, but this will break that, unfortunately.