Update patch set 1
Patch Set 1:

> Why not in USB_spec.h?

This is a Haiku-specific structure, it is not specified anywhere in the USB specification that I know of. (Linux et al. have similar structures in their API, but it is not strictly speaking something in the spec, hence why it's not in USB_spec.h.)

> Or if we create a new USB_isochronous.h, maybe move what's currently in USB_rle.h and is related to isochronous transfers in this new file?

Those structures are deprecated (they only are used by USB2.h, the R5 API, I believe.)

Patch-set: 1