Update patch set 1
Patch Set 1:

All calls I've seen used the new code, so I may be missing something.

There are still issues with face having style and drawing bits. If I save a file in StyledEdit with different bold weights and underlined text, I get just one bold weight when I reopen it. If the text is not underlined it is restored correctly, which is a small improvement to the current situation.

Patch-set: 1