Update patch set 2
Patch Set 2:

The BIOS loader is basically useless by itself. I'd imagine other loaders are the same way. The EFI loader is the exception here; so, does it really make sense to package it like this?

Why not it add it, too, to haiku_loader.hpkg? And yes, /system/boot/loaders/ seems like a better path...

Patch-set: 2
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>