Update patch set 2
Patch Set 2:

Overall, this was a half-measure to get the EFI bootloader into the installed system so users would always have access to the latest EFI loader binary.   If we needed to update the EFI loader in a stable branch, we could so so and give users some instructions on how to update it manually.

This could be useful if we shipped the native u-boot loader as well. (though less-useful with our new "efi-focused" approach)

As of your comments on packing these into haiku_loader, that would be the "ideal" method... however the MBR loader has code to parse the haiku_loader.hpkg in it and choose the "first file" as the bootloader. That code is limited to 443 bytes.

Theoretically, I tried to make sure the bios_ia32 loader was always "first" to solve this, but it really didn't work consistently.  My system booted, others systems reportedly didn't boot:

History:
https://git.haiku-os.org/haiku/commit/?id=349326db275da7f7f507540ccecb2780ac655fe9
https://git.haiku-os.org/haiku/commit/?id=92bb7e91ed4e4538a82190d313bb980262cc7057

Patch-set: 2