Update patch set 3
Patch Set 3:

WARNING: I'm not a signal processing expert

This is unnecessary for a resampling filter, where sum should be near image_filter_scale already, and certainly not in the 0 range. For other kinds of filters that may go to 0 this kind of normalization does not make sense. Not that we use any of those. [Aside: do we even use this code? I get easily lost when templates and inheritance hierarchies accumulate]

On the other hand, it is in the initialization of the weights table. I guess you'd do that once and then use the filter many times? And there are a few more conditionals in that chunk. So in this case there does not seem to be a big performance drop to care about.

Patch-set: 3
CC: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>