Update patch set 2
Patch Set 2:

Yes. I intentionally sabotaged the code in testing to ensure that the panic would actually occur in the failure condition. Outside of that, it did not seem to be triggerable for me, but I did not try a rapid unplug/replug like KapiX seems to be able to do.

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0