Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> How does it appear in https://www.haiku-os.org/docs/api/classBString.html ?

With these config changes, the constructor and the assignment operator are included in the generated documentation. It is possible to tweak the configuration in the Doxyfile to a pre-C++ 11 version, which will hide it from the documentation. This does make it possible for us to create a legacy documentation only in the future, if we wish to do so.

Patch-set: 2