Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Does this also help with #15007?

No, 15007 is a problem with the API, as explained in the ticket. Everything there is working as expected, except that the API says you can do StringWidth("aaa") + StringWidth("bbb") and get the same result as StringWidth("aaabbb"); and likewise that you can do DrawString("aaa"); DrawString("bbb") and get the same result as DrawString("aaabbb"). With hinting and subpixel rendering enabled, this isn't the case. Characters overlap and strings don't necessary align on pixel boundaries. But the value returned by StringWidth and the cursor position set by DrawString are always ceil'ed to the next pixel.

Patch-set: 1