Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > As far as I know there are no symbol names for structure fields? Offset from the struct start are hardcoded into the code?
> 
> These are static globals?

Ah, somehow I missed the changes in the cpp file (I'm too tired to be really useful today :/ )

I don't think the type is mangled into variable names. That would be useless, you can't do this anyway:

float a;
int a;

So there is no point in differenciating the two.

Patch-set: 1