Update patch set 12
Patch Set 12:

I would like to make a more meta-point. I see in the diff that you removed the original copyright holder (and license text) completely, but I wonder whether this under copyright law could be classified as original work or derivative work. Unless you rewrote the entire file from scratch, with only the man page as a reference of how it should work, (in which case I stand corrected), this feels more like a derivative work and as such should warrant keeping the original author (and/or license).

Patch-set: 12
CC: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>