Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> (1 comment)
> 
> Good point!

Actually, ... maybe I might be changing my mind. Since we are actually clearing up our own object, and reinitializing it as part of the move, it actually does not matter whether calling the destructor on this leaves the object in a valid state, as we will make it valid with the rest of the move assignment operator anyway. So there is a certain ... charm ... in disposing of the resources this uses using the destructor prior to the move. Maybe not in all cases, but at least in the case of BString.

Patch-set: 4