Update patch set 1
Patch Set 1:

> 1/ another call in vm_copy_area()
> https://git.haiku-os.org/haiku/tree/src/system/kernel/vm/vm.cpp#n2636

I will attempt to change this one.

> 2/ The free_etc call in the VMArea destructor is inconditionally with HEAP_DONT_LOCK_KERNEL_SPACE
> https://git.haiku-os.org/haiku/tree/src/system/kernel/vm/VMArea.cpp#n52

Calling free_etc() with this flag is fine, as the area will just get added to the maintenance queue and cleaned up when the deferred task runs; it won't be leaked. And, we don't really know what state we are in in the destructor, so this is OK overall. (It's used unconditionally on free_etc in other places as well.)

Actually, I see a place in the memory manager where the flag is not actually respected. That looks to be a bug.

Patch-set: 1