Update patch set 3
Patch Set 3:

BTW, we have BIcon and BIconUtils classes that implement DISABLED, SELECTED, etc. icon modifiers, you probably could use those to re-implement the shading. (I have a TODO list item to rework the BIcon class and make it do a lot of this more easily, but that's a little further out...)

Patch-set: 3