ControlLookTest: transform update rect to drawing coordinates
ControlLook methods get the size of the area for the control as well as
and update rect, a la BView::Draw, that can be use to avoid operations
on areas that will not be updated anyway. For this to be meaningful,
both rects have to be in the same coordinate system.

Change-Id: I3b08d34a24df61d1d1cbd62d296c0f126460f33e