Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Existing macros could also be undefined.
> I suppose unused functions could be removed.

I considered using undef but due to the nature of these two macros I still think renaming them is the better approach.

The unused functions seemed to me like a blueprint for continuing the implementation, particularily as the places where they are to be used were indicated. In order to maintain the spirit of the current state I just commented them out.

If there is a consensus to remove them (and the 5 corresponding call indications) I will gladly remove them.

Patch-set: 1