Update patch set 1
Patch Set 1:

I don't agree with this change.

The update rect will be computed by app_server and passed to Draw will always be in untransformed coordinates. Usually it will tell which part of the view needs to be redrawn because it was recently exposed (for example by moving another window out of the way, or by resizing the window and making i larger).

That's exactly what this test is meant to test for: that the update rect is properly handled as an untransformed one.

These conversions, if needed, should be done inside BControlLook.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>