Update patch set 1
Patch Set 1:

It was intentional to use the update rect (which is untransformed) with ConstrainClippingRegion (which is also working in untransformed coordinate space).

Clipping to the drawing rectangle is useless because nothing will be drawn outside this rectangle anyway. Clipping to the update rect is important in some cases because you don't want to overdraw existing parts of the widgets that were not cleared to some base color (for example if you have partially transparent things, like text antialiasing, re-drawing it over itself multiple times does not go well).

Probably we should document the situation about coordinate spaces, because it's not that easy to follow and with the introduction of view transforms, we now have one extra level compared to BeOS, so the be book documentation about it is not helpful anymore.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>