Update patch set 2
Patch Set 2:

Where do the previous definition of TRACE or ASSERT come from? I expect that these shouldn't be defined in public header files because we define them everywhere in a similar way, to allow to easily enable tracing for just select parts of the OS.

So it would possibly be better to remove the other definitions of the macros, instead?

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>