Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Where do the previous definition of TRACE or ASSERT come from? I expect that these shouldn't be defined in public header files because we define them everywhere in a similar way, to allow to easily enable tracing for just select parts of the OS.
> > 
> > So it would possibly be better to remove the other definitions of the macros, instead?
> 
> It was only ASSERT causing the error.
> 
> I built with: jam -q -j7 @nightly-raw
> 
> ../src/add-ons/kernel/file_systems/ext2/DirectoryIterator.cpp:26: error: "ASSERT" redefined [-Werror]
>  # define ASSERT(x) ;
>  
> In file included from ../headers/private/kernel/util/DoublyLinkedList.h:14,
>                  from ../src/add-ons/kernel/file_systems/ext2/Transaction.h:13,
>                  from ../src/add-ons/kernel/file_systems/ext2/DirectoryIterator.h:13,
>                  from ../src/add-ons/kernel/file_systems/ext2/DirectoryIterator.cpp:7:
> ../headers/private/kernel/debug.h:65: note: this is the location of the previous definition
>  # define ASSERT(x)      ASSERT_ALWAYS(x)
> 
> To me this looks like debug mechanism I would prefer not to touch.

Means I think that the original ASSERT is used for kernel debugging and used in related add-ons. And therefore I think the renaming of the ext2 macros is the way to go for the ext2 add-on.

The commenting out of the unused code depends on the state of the ext2 add-on. If it is stable and regarded as complete then the code should rather be removed. To me it looked like there was still some work to done and therefore I wanted to be least invasive. But again, I am not able decide this. 

What I am still unhappy is the cast as it exploits assumptions on the underlying data.

Patch-set: 2