Update patch set 1
Patch Set 1:

I realized this is actually the wrong way to go about things. We should really be checking this in PutUSBID instead, because if an object is being torn down to the point where it no longer can be publicly used, it should not have any children in the stack remaining either -- because those children can use it, too (e.g. DefaultPipe uses in Pipe.)

Patch-set: 1