Support Kit: Add support for advanced error types
This change introduces the BError class, and the BRuntimeError and BSystemError
subclasses. BError is an abstract base class that defines the minimum interface
for advanced error types. This includes a textual description of the error,
and a textual identifier of the origin of the error.

The error types can be used to return richer errors, either as return value or
thrown as exceptions.

Change-Id: I57723bd783b6889e9a27fb659e01696c693cf0cb