Update patch set 1
Patch Set 1: Code-Review-2

I'll have to trust you on this, as I don't even have the experience of writing a simple app and having faced any of the possible problems, but I still don't get it. Let me try to explain my view on it. Any of my points may be absurdly wrong and, unless I'm mostly right in all of them, we can abandon this and 4456 without further discussion. It's not like I want to own the drawing code, so I'm happy not knowing some details and accepting there are reasons to do things like they are done.

When the update rect comes directly from the app server, drawing is already clipped to that (except when printing?) or possibly a smaller area if there are several rects in the region, so constraining the clipping area to that update rect does nothing.

When the view is doing its stuff and wants to further restrict where the control look can draw, I say it's the caller's responsibility to clip. Using control look to draw a button would be no different than using FillEllipse to draw a circle. If I want a circle but only the part within some shape I'm drawing in, I clip to that shape and then call FillEllipse. If I want a button but only within some area, I clip to that area and then call control look to draw the button.

Thus, in my fantasy, the only use to pass an update rect is to tell the callee something like: "hey, don't even bother to calculate stuff outside this rectangle, because it will not be drawn anyway, so don't waste cycles". To do that, of course, you need the points of the clipped area and the ones you may compare them to in the same coordinate system. It looks easier to me that the transformation, if necessary, is done by the caller, who is already managing transformed coordinates anyway to tell the control look where to draw, instead of having to do it in every call in the control look.

I do agree that in HaikuControlLook clipping to the drawing rectangle is useless, except for buttons with round corners and I think sliders and maybe some other thing I have not touched yet. But it also looks made to be reused and subclassed and maybe some curly design of say _DrawFrame assumes drawing is already clipped to the widget bounds. Not that I found a guarantee like that documented anywhere, but I didn't think hard of what may be removed at this stage.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2