Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> I'll have to trust you on this, as I don't even have the experience of writing a simple app and having faced any of the possible problems, but I still don't get it. Let me try to explain my view on it. Any of my points may be absurdly wrong and, unless I'm mostly right in all of them, we can abandon this and 4456 without further discussion. It's not like I want to own the drawing code, so I'm happy not knowing some details and accepting there are reasons to do things like they are done.

There is no "right" or "wrong" here. The transforms API was added in Haiku long after the BeOS APIs were defined, and does not fit consistently with how the other APIs are used. And we never properly documented it. So, let's build a common understanding of how it should work.

Here is what I understand of how drawing works in Haiku and one possible way we need to handle these things. As far as I know, this is what is currently implemented by app_server, but there could be bugs.

> 
> When the update rect comes directly from the app server, drawing is already clipped to that (except when printing?) or possibly a smaller area if there are several rects in the region, so constraining the clipping area to that update rect does nothing.

No, it isn't. app_server will clip the drawing to the visible area of your BView (so you don't draw outside the view). But it will tell your application which part absolutely needs to be redrawn (by passing you an UpdateRect).

It is perfectly possible to draw outside of the UpdateRect by default. Even worse than that, it's possible to call drawing functions from outside Draw(), and even from another thread as long as you have the looper locked. The only clipping guaranteed by app_server is that you won't be able to draw ouside your view bounds.

Some applications will always redraw the complete view, starting by painting it with some background color and then drawing over that. Other applications will be more optimized and only redraw the part that actually needs to be redrawn.

So, there is a common use case here where you need to make sure no drawing happens outside of a given rectangle, given in non-transformed space.

If you use BDirectWindow, there is no clipping at all done by app_server, and you must be careful to do your own. However, app_server provides you with a BRegion where you are allowed to draw.

For these two cases, I think it makes sense to use GetClippingRegion/ConstrainClippingRegion.

> 
> When the view is doing its stuff and wants to further restrict where the control look can draw, I say it's the caller's responsibility to clip. Using control look to draw a button would be no different than using FillEllipse to draw a circle. If I want a circle but only the part within some shape I'm drawing in, I clip to that shape and then call FillEllipse. If I want a button but only within some area, I clip to that area and then call control look to draw the button.

If you want to do this in an application, you need to use ClipToShape or ClipToPicture (and their inverse versions). These two should follow view transforms.

> 
> Thus, in my fantasy, the only use to pass an update rect is to tell the callee something like: "hey, don't even bother to calculate stuff outside this rectangle, because it will not be drawn anyway, so don't waste cycles". To do that, of course, you need the points of the clipped area and the ones you may compare them to in the same coordinate system. It looks easier to me that the transformation, if necessary, is done by the caller, who is already managing transformed coordinates anyway to tell the control look where to draw, instead of having to do it in every call in the control look.

You cannot do this correctly, if the transform has any rotation or shearing. In these case, you cannot convert a BRect in the screen space to a BRect in the transformed space where drawing operations occurs. The result is a rotated/sheared rect, which cannot be represented by just a top left and a bottom right corner.

Moreover, you can apply transforms while you are inside BView::Draw. And you can "un-apply" them too (if you use pushstate/popstate for example). So you can't even assume the existence of a single "transformed" coordinate space during the drawing operation at all. Not to mention, what kind of coordinates should be used by the application in calls to Invalidate()? What if Invalidate() and the later corresponding Draw() call happen with the view in different transform states? Would the rectangle passed to invalidate be transformed, and then reverse-transformed again to be passed to Draw()? How would the Invalidate() rectangle be passed down to child views? How would you even check which child views are intersecting it and need to be invalidated, since child view positions are not affected by transforms?

So, unfortunately, you have to do the reverse operation: compute which pixels (in non-transformed space) are touched by your drawing, and check if that intersects with the rectangle (also in non-transformed space). Or, if that's too expensive to compute, do nothing of the sort, and let app_server handle the clipping. Again, you can use ConstrainClippingRegion here to easily have app_server do the work for you.

I don't know which side of the fence ClipToRect/ClipToInverseRect should fall on. We should just pick one and document it. The existence of an "inverse" function makes me think it should be like ClipToShape and ClipToPicture. And that also makes it consistent with FillRect/StrokeRect, jsut as ClipToShape is consistent with FillShape/StrokeShape.

So, my understanding of things is:

Things that happen in un-transformed space: Invalidate(), the updateRect passed to Draw(), and ConstrainClippingRegion().
Things that happen in transformed space: basically everything else.

For completeness we could also add a ClipToRegion that would work in transformed space (unlike ConstrainClippingRegion). I don't think there was any use for it so far.

I would like to make sure this makes sense to other people and not just me. After all, I'm not necessarily an expert here, I just happen to have been involved in the discussions at the time we added these things. It's possible that our conclusions at the time were wrong and that your approach is better. What I certainly don't want is:
- For us to end up with something that mixes both approaches and just hopes that thigns will work fine
- For these subtleties to remain undocumented

Both of these will only result in more frustration and wasted time.

Patch-set: 1